refact: change namespace for better access
