+++ samples/HttpJsonBodyParser/Voxel.MiddyNet.HttpJsonBodyParserSample/Voxel.MiddyNet.HttpJsonBodyParserSample.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Voxel.MiddyNet.HttpJsonBodyParserMiddleware\Voxel.MiddyNet.HttpJsonBodyParserMiddleware.csproj" />
+    <ProjectReference Include="..\..\..\src\Voxel.MiddyNet.HttpJsonMiddleware\Voxel.MiddyNet.HttpJsonMiddleware.csproj" />
   </ItemGroup>
 
 </Project>
