+++ samples/HttpV2JsonBodyParser/Voxel.MiddyNet.HttpV2JsonBodyParserSample/HttpV2JsonBodyParser.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Amazon.Lambda.Core;
 using Amazon.Lambda.APIGatewayEvents;
-using Voxel.MiddyNet.HttpJsonBodyParserMiddleware;
+using Voxel.MiddyNet.HttpJsonMiddleware;
 
 [assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
 namespace Voxel.MiddyNet.HttpV2JsonBodyParserSample
@@ -15,8 +15,8 @@ namespace Voxel.MiddyNet.HttpV2JsonBodyParserSample
 
         protected override Task<APIGatewayHttpApiV2ProxyResponse> Handle(APIGatewayHttpApiV2ProxyRequest lambdaEvent, MiddyNetContext context)
         {
-            var person = ((Person)context.AdditionalContext[HttpJsonBodyParserMiddleware.HttpJsonBodyParserMiddleware.BodyContextKey]);
-            context.Logger.Log(LogLevel.Info, "Function called", new LogProperty(HttpJsonBodyParserMiddleware.HttpJsonBodyParserMiddleware.BodyContextKey, person));
+            var person = ((Person)context.AdditionalContext[HttpJsonMiddleware.HttpJsonBodyParserMiddleware.BodyContextKey]);
+            context.Logger.Log(LogLevel.Info, "Function called", new LogProperty(HttpJsonMiddleware.HttpJsonBodyParserMiddleware.BodyContextKey, person));
             var result = new APIGatewayHttpApiV2ProxyResponse
             {
                 StatusCode = 200,
