+++ src/Voxel.MiddyNet.HttpJsonMiddleware/HttpJsonBodyParserMiddleware.cs
@@ -4,11 +4,15 @@ using System.Text.Json;
 using Amazon.Lambda.APIGatewayEvents;
 using System.Threading.Tasks;
 
-namespace Voxel.MiddyNet.HttpJsonBodyParserMiddleware
+namespace Voxel.MiddyNet.HttpJsonMiddleware
 {
-    public class HttpV2JsonBodyParserMiddleware<T> : HttpJsonBodyParserMiddleware, ILambdaMiddleware<APIGatewayHttpApiV2ProxyRequest, APIGatewayHttpApiV2ProxyResponse>
+    public abstract class HttpJsonBodyParserMiddleware
     {
-        public Task Before(APIGatewayHttpApiV2ProxyRequest lambdaEvent, MiddyNetContext context)
+        public const string BodyContextKey = "Body";
+    }
+    public class HttpJsonBodyParserMiddleware<T> : HttpJsonBodyParserMiddleware, ILambdaMiddleware<APIGatewayProxyRequest, APIGatewayProxyResponse> where T : new()
+    {
+        public Task Before(APIGatewayProxyRequest lambdaEvent, MiddyNetContext context)
         {
             if (!HasJsonContentHeaders(lambdaEvent))
             {
@@ -22,19 +26,19 @@ namespace Voxel.MiddyNet.HttpJsonBodyParserMiddleware
             }
 
             var source = JsonSerializer.Deserialize<T>(lambdaEvent.Body);
-            
-            
+
+
             context.AdditionalContext.Add(BodyContextKey, source);
             return Task.CompletedTask;
         }
 
-        private static bool HasJsonContentHeaders(APIGatewayHttpApiV2ProxyRequest lambdaEvent)
+        private static bool HasJsonContentHeaders(APIGatewayProxyRequest lambdaEvent)
         {
             return lambdaEvent.Headers != null &&
                    (lambdaEvent.Headers.ContainsKey("Content-Type") &&
                     lambdaEvent.Headers["Content-Type"] == "application/json");
         }
 
-        public Task<APIGatewayHttpApiV2ProxyResponse> After(APIGatewayHttpApiV2ProxyResponse lambdaResponse, MiddyNetContext context) => Task.FromResult(lambdaResponse);
+        public Task<APIGatewayProxyResponse> After(APIGatewayProxyResponse lambdaResponse, MiddyNetContext context) => Task.FromResult(lambdaResponse);
     }
 }
