+++ test/Voxel.MiddyNet.HttpJsonMiddleware.Tests/Voxel.MiddyNet.HttpJsonMiddleware.Tests.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Voxel.MiddyNet.HttpJsonBodyParserMiddleware\Voxel.MiddyNet.HttpJsonBodyParserMiddleware.csproj" />
+    <ProjectReference Include="..\..\src\Voxel.MiddyNet.HttpJsonMiddleware\Voxel.MiddyNet.HttpJsonMiddleware.csproj" />
     <ProjectReference Include="..\..\src\Voxel.MiddyNet\Voxel.MiddyNet.csproj" />
   </ItemGroup>
